requirements.txt
# Web Framework
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-CORS>=4.0.0

# Database (SQLite - no PostgreSQL needed)
SQLAlchemy>=2.0.0

# HTTP Requests
requests>=2.31.0
beautifulsoup4>=4.12.0

# AI & NLP
openai==0.28.1
numpy>=1.24.0
pandas>=2.0.0

# Environment & Configuration
python-dotenv>=1.0.0

# Utilities
python-slugify>=8.0.0
tqdm>=4.66.0

beautifulsoup4>=4.12.0
lxml>=4.9.0

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
